Moved LiveEdgeFinder, ErrorHandler, VTT and TTML Parsers into Utils to clean up the root of streaming a bit.
