+++ src/dash/DashParser.js
@@ -28,7 +28,7 @@
  *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-import ErrorHandler from '../streaming/ErrorHandler.js';
+import ErrorHandler from '../streaming/utils/ErrorHandler.js';
 import FactoryMaker from '../core/FactoryMaker.js';
 import Debug from '../core/Debug.js';
 import ObjectIron from '../../externals/objectiron.js';
