+++ src/dash/SegmentBaseLoader.js
@@ -31,7 +31,7 @@
 import RequestModifier from '../streaming/utils/RequestModifier.js';
 import Segment from './vo/Segment.js';
 import Error from '../streaming/vo/Error.js';
-import ErrorHandler from '../streaming/ErrorHandler.js';
+import ErrorHandler from '../streaming/utils/ErrorHandler.js';
 import Events from '../core/events/Events.js';
 import EventBus from '../core/EventBus.js';
 import BoxParser from '../streaming/utils/BoxParser.js';
