+++ src/streaming/MediaPlayer.js
@@ -33,8 +33,8 @@ import PlaybackController from './controllers/PlaybackController.js';
 import StreamController from './controllers/StreamController.js';
 import MediaController from './controllers/MediaController.js';
 import ManifestLoader from './ManifestLoader.js';
-import LiveEdgeFinder from './LiveEdgeFinder.js';
-import ErrorHandler from './ErrorHandler.js';
+import LiveEdgeFinder from './utils/LiveEdgeFinder.js';
+import ErrorHandler from './utils/ErrorHandler.js';
 import Capabilities from './utils/Capabilities.js';
 import TextTracks from './TextTracks.js';
 import SourceBufferController from './controllers/SourceBufferController.js';
