+++ src/streaming/controllers/SourceBufferController.js
@@ -31,11 +31,11 @@
 import TextSourceBuffer from '../TextSourceBuffer.js';
 import MediaController from './MediaController.js';
 import DashAdapter from '../../dash/DashAdapter.js';
-import ErrorHandler from '../ErrorHandler.js';
+import ErrorHandler from '../utils/ErrorHandler.js';
 import StreamController from './StreamController.js';
 import TextTracks from '../TextTracks.js';
-import VTTParser from '../VTTParser.js';
-import TTMLParser from '../TTMLParser.js';
+import VTTParser from '../utils/VTTParser.js';
+import TTMLParser from '../utils/TTMLParser.js';
 import VideoModel from '../models/VideoModel.js';
 import Error from '../vo/Error.js';
 import EventBus from '../../core/EventBus.js';
