+++ src/streaming/utils/ErrorHandler.js
@@ -28,9 +28,9 @@
  *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-import EventBus from './../core/EventBus.js';
-import Events from './../core/events/Events.js';
-import FactoryMaker from '../core/FactoryMaker.js';
+import EventBus from './../../core/EventBus.js';
+import Events from './../../core/events/Events.js';
+import FactoryMaker from '../../core/FactoryMaker.js';
 
 function ErrorHandler() {
