+++ test/dash.TimeLineConverterSpec.js
@@ -1,4 +1,4 @@
-import LiveEdgeFinder from '../src/streaming/LiveEdgeFinder.js';
+import LiveEdgeFinder from '../src/streaming/utils/LiveEdgeFinder.js';
 import EventBus from '../src/core/EventBus.js';
 import VoHelper from './helpers/VOHelper.js';
 import TimeLineConverter from '../src/dash/utils/TimelineConverter.js';
